Update _headers
Added unsafe inline to Content-Security-Policy to remedy some inline scripts and styles being blocked. Do note this is a temporary solution, as there is no user input on the website at the moment. If you want to be more secure, when user input *is* added, you'll have to remove the 'unsafe-inline' values and migrate all inline styles and scripts to external files and reference them in their place.